+++ CHANGELOG.md
@@ -54,6 +54,7 @@ The displayed date is in the format `DD-MM-YYYY`
 - feat: New cmd `reminder`
 - feat: New cmd `leaderboard`
 - feat: New cmd `backup` but it is disabled
+- feat: New cmd `autopublish` & new plugin autopublish
 
 ### Changes